# Create optimized ODataPath cloning constructor