@gathogojr 2022-08-25
Not sure I understand what you mean by this statement:
> If I make it public, a different method overload will get called without the user having to opt-in to it or without the user having to change arguments.

Are you saying that if you make this constructor public that will result into existing user code automatically invoking it instead of the constructor that is being invoked presently?