@gathogojr 2022-08-25
@habbes The following statement was confusing to me at first:
> 2. Create a new List with the same capacity as the length of the existing list

"same capacity" is what threw me off (almost). I think you should say "Create a new List with the capacity equal to the length of the existing list"