@gathogojr 2022-08-25
Under scenario 4, what does `odataPath.Segments.Length` represent on line 1 of the code snippet:
```csharp
int capacity = odataPath.Segments.Length < odataPath.segments.Count ? odataPath.segments.Capacity : odataPathSegments.Capacity + 1;
```
Did you intend:
```csharp
int capacity = odataPath.segments.Count < odataPath.segments.Capacity ? odataPath.segments.Capacity : odataPathSegments.Capacity + 1;
```