@gathogojr 2022-08-25
@habbes Curious whether there are places in the existing code that could make use of this new constructor.