@habbes 2022-08-29
> Not sure I understand what you mean by this statement:
> 
> > If I make it public, a different method overload will get called without the user having to opt-in to it or without the user having to change arguments.
> 
> Are you saying that if you make this constructor public that will result into existing user code automatically invoking it instead of the constructor that is being invoked presently?

Yes, that is precisely what I mean.

Currently we have a constructor with the existing signature:
```c#
public ODataPath(IEnumerable<ODataPathSegment>);
```
`ODataPath` also implements `IEnumerable<ODataPathSegment>`.

So we have code internally that create a copy of the `ODataPath` using the following pattern:

```c#
ODataPath oldPath;
var newPath = new ODataPath(oldPath);
```
Look at the [`AddSegment`](https://github.com/OData/odata.net/blob/7.12.2/src/Microsoft.OData.Core/UriParser/SemanticAst/ODataPathExtensions.cs#L148) extension method for example. Currently this calls the constructor above.

This PR adds a new a constructor with the signature:

```c#
internal ODataPath(ODataPath);
```
Now the `AddSegment` method's `new ODataPath(odataPath)` will go to this new constructor instead of the one that takes an `IEnumerable<ODataPath>`. But External code will still go to the old `IEnumerable<ODataPath>` constructor because the new one is internal. But if I made it public, then even external code with the same pattern will start going to the new constructor.

Note that this change won't affect calls that pass a list of segments directly, or any other enumerable for the matter. It only "hijacks" interal calls that pass an `ODataPath` directly.

Let me know if that clears things up.