@habbes 2022-08-29
> @habbes Under scenario 4, what does `odataPath.Segments.Length` represent on line 1 of the code snippet:
> 
> ```cs
> int capacity = odataPath.Segments.Length < odataPath.segments.Count ? odataPath.segments.Capacity : odataPathSegments.Capacity + 1;
> ```
> 
> Did you intend:
> 
> ```cs
> int capacity = odataPath.segments.Count < odataPath.segments.Capacity ? odataPath.segments.Capacity : odataPathSegments.Capacity + 1;
> ```
> 
> You also say that you based this PR on scenario 4 but the code in this PR is different

Yes, you're right, that is what I meant:

```cs
int capacity = odataPath.segments.Count < odataPath.segments.Capacity ? odataPath.segments.Capacity : odataPathSegments.Capacity + 1;
```

Also you're right that this slightly differs from the actual code in the PR. I must have changed after writing the description. The initial thought was to set the same capacity as the input list if was already greater than the length. But I didn't notice any improvements in the benchmarks from doing that. The thought was that doing that may reduce resizing. But that benefit didn't materialize because only one segment is going to be added to this newly created list. The next segment after that is going to be added to a whole new `ODataPath` instance with a whole new list that's a whole new copy. So it didn't make sense to optimize beyond the immediate next segment to be added, otherwise we may just end up with a bunch of arrays that are larger than necessary. Let me update the description to match the code.