@habbes 2022-08-29
> @habbes Curious whether there are places in the existing code that could make use of this new constructor.

Yes, I have mentiond an example in a comment above: https://github.com/OData/odata.net/pull/2489#issuecomment-1229753241