@habbes 2022-08-29
> @habbes I'm not sure how much we do equality checks for `ODataPath` but maybe we could also optimize by using a for loop in place of `Where` and `Any` [here]that (
> 
> https://github.com/OData/odata.net/blob/4e263ca4b94d01f2d105d8282c62ec272cfcd1bb/src/Microsoft.OData.Core/UriParser/SemanticAst/ODataPath.cs#L152
> 
> )

Yeah, we could make that improvement. But I have seen a couple of other places that could benefit from using plain loops to iterate over the ODataPath. I think I could bundle them up in a separate PR, what do you think?